Uploading all the last changes

+++ classrooms/school_users/views.py
@@ -1270,14 +1270,18 @@ def delete_user(request, user_id=None, type_of_user=None):
             user_to_delete = get_object_or_404(Parent, parent_id=user_id)
         if(type_of_user=='student'):
             user_to_delete = get_object_or_404(Student, student_id=user_id)
-            if user_to_delete.first_parent:
-                if user_to_delete.first_parent.profile:
-                    user_to_delete.first_parent.profile.delete()
-                user_to_delete.first_parent.delete()
-            if user_to_delete.second_parent:
-                if user_to_delete.second_parent.profile:
-                    user_to_delete.second_parent.profile.delete()
-                user_to_delete.second_parent.delete()
+            if user_to_delete.first_parent and user_to_delete.second_parent:
+                if not user_to_delete.first_parent == user_to_delete.second_parent:
+                    if user_to_delete.first_parent.profile:
+                        user_to_delete.first_parent.profile.delete()
+                    user_to_delete.first_parent.delete()
+                    if user_to_delete.second_parent.profile:
+                        user_to_delete.second_parent.profile.delete()
+                    user_to_delete.second_parent.delete()
+                else:
+                    if user_to_delete.first_parent.profile:
+                        user_to_delete.first_parent.profile.delete()
+                    user_to_delete.first_parent.delete()
         if user_to_delete.profile:
             user_to_delete.profile.delete()
         user_to_delete.delete()
